src/backend/auth/pyproject.toml
[project]
name = "auth"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.18.3",
    "argon2-cffi>=25.1.0",
    "asyncpg==0.31.0",
    "blake3>=1.0.8",
    "cashews[redis]==7.4.4",
    "greenlet>=3.3.1",
    "grpcio==1.76.0",
    "grpcio-tools==1.76.0",
    "httpagentparser==1.9.5",
    "jwskate==0.12.2",
    "py-async-grpc-prometheus==0.2.0",
    "sqlalchemy==2.0.46",
    "uvicorn[standard]==0.40.0",
]

[dependency-groups]
test = [
    "coverage==7.13.3",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
]
